chore (Modules) Remove from the DOM iframes which are not visible
When navigating from Tab A to Tab B, hide iframes on Tab A and
present iframes on Tab B. Also, on layers.push(), hide iframes
that were on the bottom layer and bring them back on layers.pop().

This allows to save device memory since the rendering context can
be kept smaller.